Fixed a "bug" on 2nd stage of Level 5 where player has to even the scales
Fixed a "bug" on 2nd stage of Level 5 where player has to even the scales. Now player can place gold on either side and bronze/silver on the remaining side. Before fixing the "bug" it was possible only to even the scales by placing gold on the right side and silver/bronze on the left side.